+++ CHANGELOG.md
@@ -42,6 +42,8 @@ From v2.1.0 to v3.0.0
   after fitting.
 - Tutorials were completely rewritten and expanded. They can now be found in
   the `tutorials` folder instead of the `tutorial` folder.
+- `MetricsPlot` now supports multi-class and multi-task classification models.
+- `CorrelationPlot` now supports multi-task regression models.
 
 ## New Features
 
@@ -69,6 +71,7 @@ From v2.1.0 to v3.0.0
   showcasing these features was created.
 - Added a `reset` method to `QSPRDataset`, which resets splits and loads all descriptors
   into the training set matrix again.
+- Added `ConfusionMatrixPlot` to plot confusion matrices.
 
 ## Removed Features